+++ probability_version.py
@@ -1,2 +1,2 @@
-VERSION = "0.6.0"
+VERSION = "0.6.2"
 APP_NAME = "Probability Platform"