+++ tests/test_version.py
@@ -1,4 +1,4 @@
 from probability_version import VERSION
 
 def test_version():
-    assert VERSION == "0.6.0"
+    assert VERSION == "0.6.2"